refactor(backend): move project crud into its own module
Pure move: every function body is byte-identical to its original in
crud/__init__.py, verified by parsing both files with ast. delete_project
moves back next to its siblings instead of sitting in the test_result block.

A transitional re-export at the extraction site keeps callers working so this
commit is reviewable on its own; the next commit migrates them and drops it.
It sits at the extraction site rather than in the import block so the import
header stays byte-identical to main and cannot conflict with #2410 or #2411.
